WIP on exporting attachment as yaml
