+++ internal/ext/importer.go
@@ -1,0 +1,154 @@
+package ext
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+
+	"github.com/markphelps/flipt/rpc/flipt"
+	"github.com/markphelps/flipt/storage"
+	"gopkg.in/yaml.v2"
+)
+
+type Importer struct {
+	ctx   context.Context
+	store storage.Store
+}
+
+func NewImporter(ctx context.Context, store storage.Store) *Importer {
+	return &Importer{
+		ctx:   ctx,
+		store: store,
+	}
+}
+
+func (i *Importer) ReadFrom(r io.Reader) (int64, error) {
+	var (
+		ctx = i.ctx
+		dec = yaml.NewDecoder(r)
+		doc = new(Document)
+	)
+
+	if err := dec.Decode(doc); err != nil {
+		return -1, fmt.Errorf("importing: %w", err)
+	}
+
+	var (
+		// map flagKey => *flag
+		createdFlags = make(map[string]*flipt.Flag)
+		// map segmentKey => *segment
+		createdSegments = make(map[string]*flipt.Segment)
+		// map flagKey:variantKey => *variant
+		createdVariants = make(map[string]*flipt.Variant)
+	)
+
+	// create flags/variants
+	for _, f := range doc.Flags {
+		flag, err := i.store.CreateFlag(ctx, &flipt.CreateFlagRequest{
+			Key:         f.Key,
+			Name:        f.Name,
+			Description: f.Description,
+			Enabled:     f.Enabled,
+		})
+
+		if err != nil {
+			return -1, fmt.Errorf("importing flag: %w", err)
+		}
+
+		for _, v := range f.Variants {
+			var (
+				out []byte
+				err error
+			)
+
+			if v.Attachment != nil {
+				out, err = json.Marshal(v.Attachment)
+				if err != nil {
+					return -1, fmt.Errorf("importing flag: invalid variant attachment type: %T", v.Attachment)
+				}
+			}
+
+			variant, err := i.store.CreateVariant(ctx, &flipt.CreateVariantRequest{
+				FlagKey:     f.Key,
+				Key:         v.Key,
+				Name:        v.Name,
+				Description: v.Description,
+				Attachment:  string(out),
+			})
+
+			if err != nil {
+				return -1, fmt.Errorf("importing variant: %w", err)
+			}
+
+			createdVariants[fmt.Sprintf("%s:%s", flag.Key, variant.Key)] = variant
+		}
+
+		createdFlags[flag.Key] = flag
+	}
+
+	// create segments/constraints
+	for _, s := range doc.Segments {
+		segment, err := i.store.CreateSegment(ctx, &flipt.CreateSegmentRequest{
+			Key:         s.Key,
+			Name:        s.Name,
+			Description: s.Description,
+		})
+
+		if err != nil {
+			return -1, fmt.Errorf("importing segment: %w", err)
+		}
+
+		for _, c := range s.Constraints {
+			_, err := i.store.CreateConstraint(ctx, &flipt.CreateConstraintRequest{
+				SegmentKey: s.Key,
+				Type:       flipt.ComparisonType(flipt.ComparisonType_value[c.Type]),
+				Property:   c.Property,
+				Operator:   c.Operator,
+				Value:      c.Value,
+			})
+
+			if err != nil {
+				return -1, fmt.Errorf("importing constraint: %w", err)
+			}
+		}
+
+		createdSegments[segment.Key] = segment
+	}
+
+	// create rules/distributions
+	for _, f := range doc.Flags {
+		// loop through rules
+		for _, r := range f.Rules {
+			rule, err := i.store.CreateRule(ctx, &flipt.CreateRuleRequest{
+				FlagKey:    f.Key,
+				SegmentKey: r.SegmentKey,
+				Rank:       int32(r.Rank),
+			})
+
+			if err != nil {
+				return -1, fmt.Errorf("importing rule: %w", err)
+			}
+
+			for _, d := range r.Distributions {
+				variant, found := createdVariants[fmt.Sprintf("%s:%s", f.Key, d.VariantKey)]
+				if !found {
+					return -1, fmt.Errorf("finding variant: %s; flag: %s", d.VariantKey, f.Key)
+				}
+
+				_, err := i.store.CreateDistribution(ctx, &flipt.CreateDistributionRequest{
+					FlagKey:   f.Key,
+					RuleId:    rule.Id,
+					VariantId: variant.Id,
+					Rollout:   d.Rollout,
+				})
+
+				if err != nil {
+					return -1, fmt.Errorf("importing distribution: %w", err)
+				}
+			}
+		}
+	}
+
+	return 0, nil
+}
